requirements.txt
certifi==2021.5.30
click>=8.1.3
Flask==2.3.3
gunicorn>=20.1.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.1
Werkzeug>=1.0.1
wincertstore==0.2
opencv-python-headless==4.6.0.66
numpy>=1.23.1
matplotlib==3.5.3
requests==2.28.1
tensorflow-cpu>=1.21.6
transformers==4.33.3